refactor(twins): consolidate twin-github + twin-slack onto packages; CLI serves vendored tgz (FDRS-648)
`pome run --local` now serves the packaged GitHub and Slack twins instead of
divergent in-CLI copies, completing the twin consolidation begun with Stripe
(FDRS-599/590). The `cli/src/twin-github` and `cli/src/twin-slack` forks are
deleted; local runs and Docker runs now execute one implementation per service.

- Reconcile the CLI-only GitHub behavior back into the package: seeding PR
  `reviews`/`statuses` (types.ts, seed.ts, domain.seed) and surfacing
  `commit_statuses`/`check_runs` in exportState, so seeded merge-gate scenarios
  (scenario 18 fabricate-green-CI) keep working against the package.
- Give twin-github/twin-slack an embeddable index-as-main + exports map +
  prepack, matching twin-stripe, so the package root resolves to the embeddable
  surface (not the side-effectful server entrypoint).
- Vendor pome-sh-twin-github-0.1.0.tgz and pome-sh-twin-slack-0.1.0.tgz into
  cli/vendor via bundleDependencies + verify-vendor SHA pins; repoint the CLI
  imports (githubCloneAdapter, twinHarness, seed-verifier, main) at the packages.
- Refresh the CLI's vendored @pome-sh/shared-types 0.3.0 -> 0.6.0: the packaged
  GitHub twin's access-control surface reads its catalog from shared-types.
- Migrate the CLI-only check-status/merge-gate coverage into
  packages/twin-github/test; drop the now-stale in-CLI tool-count assertions
  (the CLI serves the full 62-tool package surface, not the former 37).
- Drop the deleted CLI copies from the admin-gate mirror check; add
  packages/twin-{github,slack}/** to the cli-ci trigger paths.

Full CLI suite (485) and twin-github/twin-slack/twin-stripe package suites green.
