+++ cli/scripts/verify-vendor.mjs
@@ -7,7 +7,9 @@ import { fileURLToPath } from "node:url";
 const root = resolve(dirname(fileURLToPath(import.meta.url)), "..");
 
 const EXPECTED_SHA256 = new Map([
-  ["vendor/pome-sh-shared-types-0.3.0.tgz", "284ef36d0b8c138a3d66998916655962da8452ccd6882b3f2dfaf461e5c7a1b9"],
+  ["vendor/pome-sh-shared-types-0.6.0.tgz", "81ac63582f8e27e86bfec0b72476b470f90b2ef894b3ea6b9df8c581d9b4eaf4"],
+  ["vendor/pome-sh-twin-github-0.1.0.tgz", "fb09bf8858465a1efbf9d4d5922592afc87dd0260de07ab6e11d59f3e89e622f"],
+  ["vendor/pome-sh-twin-slack-0.1.0.tgz", "d212c6fb03c1de0f6f636018b0b643b273846d3bcf98e969b67bf6dd67eaf648"],
   ["vendor/pome-sh-twin-stripe-0.2.0.tgz", "bed2fdc0d6c92921079d21400fb60b0569897a124cf48e52fd9b0e1fbde76e5f"],
 ]);
