+++ cli/test/integration/githubCloneAdapter.test.ts
@@ -30,7 +30,9 @@ describe("github_clone adapter", () => {
       headers: { Authorization: `Bearer ${token}` }
     });
     const tools = (await toolsResponse.json()) as { tools: Array<{ name: string }> };
-    expect(tools.tools).toHaveLength(37);
+    // Full packaged twin-github surface (consolidated in FDRS-648): the CLI now
+    // serves @pome-sh/twin-github, not the former in-CLI subset (was 37).
+    expect(tools.tools).toHaveLength(62);
     const toolNames = tools.tools.map((tool) => tool.name);
     expect(toolNames).toContain("merge_pull_request");
     expect(toolNames).toContain("create_commit_status");
@@ -63,14 +65,14 @@ describe("github_clone adapter", () => {
     });
     expect(notifRes.status).toBe(202);
 
-    // tools/list → 37 tools with camelCase inputSchema
+    // tools/list → full packaged twin-github surface with camelCase inputSchema
     const listRes = await app.request(url, {
       method: "POST",
       headers,
       body: JSON.stringify({ jsonrpc: "2.0", id: 2, method: "tools/list" })
     });
     const listJson = (await listRes.json()) as { result: { tools: Array<{ name: string; inputSchema: unknown }> } };
-    expect(listJson.result.tools).toHaveLength(37);
+    expect(listJson.result.tools).toHaveLength(62);
     expect(listJson.result.tools[0]).toHaveProperty("inputSchema");
     expect(listJson.result.tools[0]).not.toHaveProperty("input_schema");
